Condense some code, deciding where to take this application in the future.
It's become apparent to me that this application has the potential to be something more than a simple time clock application. Maybe I should turn it into an HR monster machine, where it calculates pay rates, taxes, and other things based on a User's roles. Maybe figure out how to deploy the application on multiple devices with a home hub for a manager to do admin stuff on. Maybe make it call home to a server, and multiple organizations can access it for themselves, and give admin rights to certain users per organization. The possibilities are endless!